[dmitry]:
r524
Controller fsCakeWashingControl is modified accordingly to the convention.
---------------------
r523
fsCakeWashingCalculator.cs is added (with appropriately modified parameters accordingly to convention)
---------------------
r522
Comments for the equations in Cake Washing folder are added and class Equation is renamed where it needed to be done.
---------------------
r513
Sorry. In the previous commit the folder Cake Washing hadn't be added. Now everything is ok!
---------------------
r512
The function Erfc is added to the file fsValue.cs. The folder Cake Washing is created and the set of main Cake Washing Equations are settled in this folder.
---------------------
r508
File fsParameterIdentifier.cs is updated. New parameters are added. 
---------------------
r507
File fsCharacteristic.cs is updated. New characteristics are added.
---------------------
r506
The file fsUnits.cs is modified (the new units are added)
---------------------